fix: gate CUDA resize filters to Linux
